geocode an address from the same form
add geocode to attributes
add button to template
add button event-handler
the geocode functionality searches for fields named "address.street", "address.zip" and "address.city" in the current form, concatenates the values and tries to geocode them using google.maps.Geocoder. In case a result / location is returned it is used as the value of our field, the marker position and map center are updated.
TODO: find a cleaner solution to get the address or combine address inputs and map in a new autoform input.